+++ src/Math/BigIntCalculator.cs
@@ -66,7 +66,7 @@ public sealed class BigIntCalculator : Calculator<BigInteger>
     public override bool Equals(Calculator<BigInteger> other)
     {
         var valueLeft = this.Value.ToByteArray();
-        var valueRight = other?.Value.ToByteArray() ?? Array.Empty<byte>();
+        var valueRight = other?.Value.ToByteArray() ?? [];
 
         var diff = (uint)valueLeft.Length ^ (uint)valueRight.Length;
         for (var i = 0; i < valueLeft.Length && i < valueRight.Length; i++)
@@ -82,28 +82,28 @@ public sealed class BigIntCalculator : Calculator<BigInteger>
     /// </summary>
     /// <param name="right">right-hand operand</param>
     /// <returns>This method returns <see langword="true"/> if this instance is greater than the <paramref name="right"/> instance, <see langword="false"/> otherwise.</returns>
-    public override bool GreaterThan(BigInteger right) => this.Value > right;
+    protected override bool GreaterThan(BigInteger right) => this.Value > right;
 
     /// <summary>
     /// This method represents the Lower Than operator.
     /// </summary>
     /// <param name="right">right-hand operand</param>
     /// <returns>This method returns <see langword="true"/> if this instance is less than the <paramref name="right"/> instance, <see langword="false"/> otherwise.</returns>
-    public override bool LowerThan(BigInteger right) => this.Value < right;
+    protected override bool LowerThan(BigInteger right) => this.Value < right;
 
     /// <summary>
     /// This method represents the Greater Than Or Equal To operator.
     /// </summary>
     /// <param name="right">right-hand operand</param>
     /// <returns>This method returns <see langword="true"/> if this instance is greater than or equal to the <paramref name="right"/> instance, <see langword="false"/> otherwise.</returns>
-    public override bool EqualOrGreaterThan(BigInteger right) => this.Value >= right;
+    protected override bool EqualOrGreaterThan(BigInteger right) => this.Value >= right;
 
     /// <summary>
     /// This method represents the Lower Than Or Equal To operator.
     /// </summary>
     /// <param name="right">right-hand operand</param>
     /// <returns>This method returns <see langword="true"/> if this instance is less than or equal to the <paramref name="right"/> instance, <see langword="false"/> otherwise.</returns>
-    public override bool EqualOrLowerThan(BigInteger right) => this.Value <= right;
+    protected override bool EqualOrLowerThan(BigInteger right) => this.Value <= right;
 
     /// <inheritdoc />
     /// <exception cref="T:System.OverflowException">Unable to convert the current instance of <see cref="BigIntCalculator"/> class to <see cref="Int32"/>.</exception>
@@ -127,7 +127,7 @@ public sealed class BigIntCalculator : Calculator<BigInteger>
     /// <param name="right">Right value to add (right summand).</param>
     /// <returns>The sum of the current <see cref="BigIntCalculator"/> instance and the <paramref name="right"/>
     /// <see cref="BigIntCalculator"/> instance.</returns>
-    public override Calculator<BigInteger> Add(BigInteger right) => this.Value + right;
+    protected override Calculator<BigInteger> Add(BigInteger right) => this.Value + right;
 
     /// <summary>
     /// Subtracts the current <see cref="BigIntCalculator"/> instance with the <paramref name="right"/>
@@ -136,7 +136,7 @@ public sealed class BigIntCalculator : Calculator<BigInteger>
     /// <param name="right">Right value to subtract (subtrahend).</param>
     /// <returns>The difference of the current <see cref="BigIntCalculator"/> instance and the <paramref name="right"/>
     /// <see cref="BigIntCalculator"/> instance.</returns>
-    public override Calculator<BigInteger> Subtract(BigInteger right) => this.Value - right;
+    protected override Calculator<BigInteger> Subtract(BigInteger right) => this.Value - right;
 
     /// <summary>
     /// Multiplies the current <see cref="BigIntCalculator"/> instance with the <paramref name="right"/>
@@ -145,7 +145,7 @@ public sealed class BigIntCalculator : Calculator<BigInteger>
     /// <param name="right">multiplicand</param>
     /// <returns>The product of the current <see cref="BigIntCalculator"/> instance and the <paramref name="right"/>
     /// <see cref="BigIntCalculator"/> instance.</returns>
-    public override Calculator<BigInteger> Multiply(BigInteger right) => this.Value * right;
+    protected override Calculator<BigInteger> Multiply(BigInteger right) => this.Value * right;
 
     /// <summary>
     /// Divides the current <see cref="BigIntCalculator"/> instance with the <paramref name="right"/>
@@ -154,26 +154,26 @@ public sealed class BigIntCalculator : Calculator<BigInteger>
     /// <param name="right">divisor</param>
     /// <returns>The quotient of the current <see cref="BigIntCalculator"/> instance and the <paramref name="right"/>
     /// <see cref="BigIntCalculator"/> instance.</returns>
-    public override Calculator<BigInteger> Divide(BigInteger right) => this.Value / right;
+    protected override Calculator<BigInteger> Divide(BigInteger right) => this.Value / right;
 
     /// <summary>
     /// The modulo operation
     /// </summary>
     /// <param name="right">divisor</param>
     /// <returns>The remainder as <see cref="BigIntCalculator"/> instance.</returns>
-    public override Calculator<BigInteger> Modulo(BigInteger right) => this.Value % right;
+    protected override Calculator<BigInteger> Modulo(BigInteger right) => this.Value % right;
 
     /// <summary>
     /// The unary increment method increments this instance by 1.
     /// </summary>
     /// <returns>This <see cref="BigIntCalculator"/> instance plus <see cref="Calculator{BigInteger}.One"/></returns>
-    public override Calculator<BigInteger> Increment() => ++this.Clone().Value;
+    protected override Calculator<BigInteger> Increment() => ++this.Clone().Value;
 
     /// <summary>
     /// The unary decrement method decrements this instance by 1.
     /// </summary>
     /// <returns>This <see cref="BigIntCalculator"/> instance minus <see cref="Calculator{BigInteger}.One"/></returns>
-    public override Calculator<BigInteger> Decrement() => --this.Clone().Value;
+    protected override Calculator<BigInteger> Decrement() => --this.Clone().Value;
 
     /// <summary>
     /// Returns the absolute value of the current <see cref="BigIntCalculator"/> object.
@@ -202,17 +202,17 @@ public sealed class BigIntCalculator : Calculator<BigInteger>
     public override IEnumerable<byte> ByteRepresentation => new ReadOnlyCollection<byte>(this.Value.ToByteArray());
 
     /// <summary>
-    /// Gets a value indicating whether or not the current <see cref="BigIntCalculator"/> object is zero (0).
+    /// Gets a value indicating whether the current <see cref="BigIntCalculator"/> object is zero (0).
     /// </summary>
     public override bool IsZero => this.Value.IsZero;
 
     /// <summary>
-    /// Gets a value indicating whether or not the current <see cref="BigIntCalculator"/> object is one (1).
+    /// Gets a value indicating whether the current <see cref="BigIntCalculator"/> object is one (1).
     /// </summary>
     public override bool IsOne => this.Value.IsOne;
 
     /// <summary>
-    /// Gets a value indicating whether or not the current <see cref="BigIntCalculator"/> object is an even number.
+    /// Gets a value indicating whether the current <see cref="BigIntCalculator"/> object is an even number.
     /// </summary>
     public override bool IsEven => this.Value % 2 == 0;
 
@@ -225,30 +225,27 @@ public sealed class BigIntCalculator : Calculator<BigInteger>
     /// Returns the square root of the current <see cref="BigIntCalculator"/> object.
     /// </summary>
     /// <exception cref="T:System.ArithmeticException" accessor="get">NaN (value is lower than zero)</exception>
-    public override Calculator<BigInteger> Sqrt
+    public override Calculator<BigInteger> Sqrt()
     {
-        get
+        if (this.Value == BigInteger.Zero)
         {
-            if (this.Value == BigInteger.Zero)
-            {
-                return Zero;
-            }
-
-            if (this.Value < BigInteger.Zero)
-            {
-                throw new ArithmeticException("NaN");
-            }
-
-            int bitLength = Convert.ToInt32(Math.Ceiling(BigInteger.Log(this.Value, 2)));
-            var root = BigInteger.One << (bitLength >> 1);
-            bool IsSqrt(BigInteger n, BigInteger r) => n >= r * r && n < (r + 1) * (r + 1);
-            while (!IsSqrt(this.Value, root))
-            {
-                root = root + this.Value / root >> 1;
-            }
-
-            return root;
+            return Zero;
         }
+
+        if (this.Value < BigInteger.Zero)
+        {
+            throw new ArithmeticException("NaN");
+        }
+
+        int bitLength = Convert.ToInt32(Math.Ceiling(BigInteger.Log(this.Value, 2)));
+        var root = BigInteger.One << (bitLength >> 1);
+        bool IsSqrt(BigInteger n, BigInteger r) => n >= r * r && n < (r + 1) * (r + 1);
+        while (!IsSqrt(this.Value, root))
+        {
+            root = root + this.Value / root >> 1;
+        }
+
+        return root;
     }
 
     /// <summary>